DataType file re-org
lib/familia/data_type.rb
lib/familia/data_type/
├── class_methods.rb      # Class-level methods for DataType itself
├── settings.rb           # Instance-level settings/configuration
├── connection.rb         # Instance-level connection & key generation
├── commands.rb           # Instance-level Redis commands (updated)
├── serialization.rb      # Instance-level value serialization (updated)
└── types/                # Concrete implementations (UNCHANGED)
